- Took the transaction-payment pallet back out - Put the signed extension  of the transaction-payment pallet in, so we can now hard-code a transaction fee.
