--- pallets/transaction-payment/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "pallet-transaction-payment"
-version = "1.0.1"
-authors = ["Parallax"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/substrate/"
-description = "FRAME pallet to manage transaction payments"
-readme = "README.md"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.1", default-features = false, features = ["derive"] }
-serde = { version = "1.0.101", optional = true }
-sp-std = { version = "2.0.0", default-features = false }
-sp-runtime = { version = "2.0.0", default-features = false}
-frame-support = { version = "2.0.0", default-features = false }
-frame-system = { version = "2.0.0", default-features = false }
-pallet-transaction-payment-rpc-runtime-api = { version = "2.0.0", default-features = false }
-smallvec = "1.4.1"
-sp-io = { version = "2.0.0", default-features = false }
-sp-core = { version = "2.0.0", default-features = false }
-
-[dev-dependencies]
-pallet-balances = { version = "2.0.0" }
-sp-storage = { version = "2.0.0" }
-
-[features]
-default = ["std"]
-std = [
-	"serde",
-	"codec/std",
-	"sp-std/std",
-	"sp-runtime/std",
-	"frame-support/std",
-	"frame-system/std",
-	"pallet-transaction-payment-rpc-runtime-api/std",
-	"sp-io/std",
-	"sp-core/std",
-]
