+++ runtime/src/lib.rs
@@ -352,7 +352,7 @@ pub type SignedExtra = (
 	frame_system::CheckEra<Runtime>,
 	frame_system::CheckNonce<Runtime>,
 	frame_system::CheckWeight<Runtime>,
-	pallet_transaction_payment::ChargeTransactionPayment<Runtime>,
+	pallet_subtensor::ChargeTransactionPayment<Runtime>,
 	pallet_subtensor::FeeFromSelfEmission<Runtime>
 );
 /// Unchecked extrinsic type as expected by this runtime.
